Changed GameElement from trait to abstract class to allow the excess Listbuffer in the cave  to be of type GameElement rather than of AnyRef. This allowed it to searched by Index, Name, and Type in the search function and allowed it's results to be shown.
Changed the search function when searching by Index so it will now convert the index of the elements in the cave to a string rather than attempt to convert the input of the search into an int. This will allow for the app to properly handle improper inputs.

In the functions that add the game elements to their proper place in the cave, I moved where the function checks to see if the element is not a part of any party or creature. It was inadvertently located in the foreach loop and this caused multiple duplicate entries to be added to the excess listbuffer.